作业四 集成CUDA

+++ src/ops/rms_norm/cpu/rms_norm_cpu.cpp
@@ -32,7 +32,7 @@ void rms_norm_impl(
 
         // 计算均方根倒数，避免在下面的循环中多次除法运算
         float mean_of_squares = sum_of_squares / static_cast<float>(norm_size);
-        float inv_rms = 1.0f / std::sqrt(mean_of_squares + eps);
+        float inv_rms = 1.0f / sqrtf(mean_of_squares + eps);
 
         // 归一化，并乘以权重
         for (size_t i =0; i < norm_size; ++i) {
